// src/enum/product-category.enum.ts
export enum ProductCategory {
    TrvanlivePotraviny = 'trvanlivePotraviny',
    SpecialnaAZdravaVyziva = 'specialnaAZdravaVyziva',
    Pecivo = 'pecivo',
    OvocieAZeleniny = 'ovocieAZeleniny',
    Napoje = 'napoje',
    MrazenePotraviny = 'mrazenePotraviny',
    MliecneVyrobkyAVajcia = 'mliecneVyrobkyAVajcia',
    MasoRybyALahodky = 'masoRybyALahodky',
    Grilovanie = 'grilovanie',
    Alkohol = 'alkohol',
}